# cf/openapi.yaml
openapi: 3.0.0
info:
  title: OpenMeter Cloudflare Workers API
  version: 1.0.0
  description: |
    OpenMeter API implementation using Cloudflare Workers, D1, and KV.

    This API provides usage metering and billing capabilities with:
    - Real-time event ingestion
    - Flexible meter configurations
    - Subject and feature management
    - Usage analytics and reporting
  contact:
    name: OpenMeter Team
    url: https://openmeter.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://your-worker.your-subdomain.workers.dev
    description: Production server
  - url: http://localhost:8787
    description: Development server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /api/v1/meters:
    get:
      tags: [Meters]
      summary: List meters
      description: Get a paginated list of meters
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Search"
        - name: aggregation
          in: query
          description: Filter by aggregation type
          schema:
            $ref: "#/components/schemas/MeterAggregation"
      responses:
        "200":
          description: List of meters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedMeterResponse"

    post:
      tags: [Meters]
      summary: Create meter
      description: Create a new meter configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMeterRequest"
      responses:
        "201":
          description: Meter created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meter"
        "409":
          description: Meter with key already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/meters/{id}:
    get:
      tags: [Meters]
      summary: Get meter
      description: Get a meter by ID
      parameters:
        - $ref: "#/components/parameters/MeterId"
      responses:
        "200":
          description: Meter details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meter"
        "404":
          description: Meter not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags: [Meters]
      summary: Update meter
      description: Update an existing meter (admin only)
      parameters:
        - $ref: "#/components/parameters/MeterId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeterRequest"
      responses:
        "200":
          description: Meter updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meter"
        "404":
          description: Meter not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Meters]
      summary: Delete meter
      description: Soft delete a meter (admin only)
      parameters:
        - $ref: "#/components/parameters/MeterId"
      responses:
        "200":
          description: Meter deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  id:
                    type: string
                  deletedAt:
                    type: string
                    format: date-time
...